flask-server/requirements.txt
flask
python-dotenv
openai
langchain
pandas
numpy

pypdf2 
docx2txt 
faiss-cpu 
tiktoken

pymongo